fixed boot receivers and added delete confirmation dialog

+++ app/src/main/java/com/ulan/timetable/adapters/ExamsAdapter.java
@@ -121,10 +121,12 @@ public class ExamsAdapter extends ArrayAdapter<Exam> {
                     public boolean onMenuItemClick(@NonNull MenuItem item) {
                         int itemId = item.getItemId();
                         if (itemId == R.id.delete_popup) {
-                            db.deleteExamById(getItem(position));
-                            db.updateExam(getItem(position));
-                            examlist.remove(position);
-                            notifyDataSetChanged();
+                            AlertDialogsHelper.getDeleteDialog(getContext(), () -> {
+                                db.deleteExamById(getItem(position));
+                                db.updateExam(getItem(position));
+                                examlist.remove(position);
+                                notifyDataSetChanged();
+                            });
                             return true;
                         } else if (itemId == R.id.edit_popup) {
                             final View alertLayout = mActivity.getLayoutInflater().inflate(R.layout.timetable_dialog_add_exam, null);
